requirements.txt
confluent-kafka==1.1.0
pytest-docker-compose==3.1.2
